+++ parser/src/parser_logic.rs
@@ -1,5 +1,6 @@
 use super::lang;
-use program_structure::ast::{AST, produce_report, produce_generic_report};
+use program_structure::ast::{AST};
+use program_structure::error_code::{produce_report, produce_generic_report};
 use program_structure::error_code::ReportCode;
 use program_structure::error_definition::{ReportCollection};
 use program_structure::file_definition::FileID;
