+++ program_structure/src/program_library/error_code.rs
@@ -1,6 +1,8 @@
 use core::fmt;
 use std::fmt::Formatter;
 
+use crate::{error_definition::Report, file_definition::{FileLocation, FileID}, ast::Version};
+
 #[derive(Copy, Clone)]
 pub enum ReportCode {
     //Parse Errors
@@ -206,3 +208,101 @@ impl fmt::Display for ReportCode {
         f.write_str(string_format)
     }
 }
+
+
+
+
+pub fn produce_report_with_message(error_code : ReportCode, msg : String) -> Report {
+    match error_code {
+        ReportCode::FileOs => {
+            Report::error(
+            format!("Could not open file {}", msg),
+            ReportCode::FileOs,
+            )
+        }
+        ReportCode::IncludeNotFound => {
+            Report::error(
+                format!(" The file {} to be included has not been found", msg),
+                ReportCode::FileOs,
+                )
+        },
+        _ => unreachable!()
+    }
+}
+
+pub fn produce_generic_report( msg : String, location : FileLocation, file_id : FileID) -> Report {
+    let mut report = Report::error(msg, ReportCode::GenericParsing);
+    report.add_primary(location, file_id, "Error here".to_string());
+    report
+}
+pub fn produce_compiler_version_report(path : String, required_version : Version, version :  Version) -> Report {
+    let report = Report::error(
+        format!("File {} requires pragma version {:?} that is not supported by the compiler (version {:?})", path, required_version, version ),
+        ReportCode::CompilerVersionError,
+    );
+    report
+}
+
+
+ pub fn produce_report(error_code: ReportCode, location : FileLocation, file_id : FileID) -> Report {
+    use ReportCode::*;
+    let report  = match error_code {
+            UnclosedComment => {
+                let mut report =
+                    Report::error("unterminated /* */".to_string(), ReportCode::UnclosedComment);
+                report.add_primary(location, file_id, "Comment starts here".to_string());
+                report
+            }
+            NoMain => Report::error(
+                "No main specified in the project structure".to_string(),
+                ReportCode::NoMainFoundInProject,
+            ),
+            MultipleMain =>{
+                Report::error(
+                    "Multiple main components in the project structure".to_string(),
+                    ReportCode::MultipleMainInComponent,
+                )
+            }
+            MissingSemicolon => {
+                let mut report = Report::error(format!("Missing semicolon"), 
+                    ReportCode::MissingSemicolon);
+                report.add_primary(location, file_id, "A semicolon is needed here".to_string());
+                report
+            }
+            UnrecognizedInclude => {
+                let mut report =
+                Report::error("unrecognized argument in include directive".to_string(), ReportCode::UnrecognizedInclude);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+
+            }
+            UnrecognizedPragma => {
+                let mut report =
+                Report::error("unrecognized argument in pragma directive".to_string(), ReportCode::UnrecognizedPragma);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+
+            }        
+            UnrecognizedVersion => {
+                let mut report =
+                Report::error("unrecognized version argument in pragma directive".to_string(), ReportCode::UnrecognizedVersion);
+            report.add_primary(location, file_id, "this argument".to_string());
+            report
+            }      
+            IllegalExpression => {
+                let mut report =
+                Report::error("illegal expression".to_string(), ReportCode::IllegalExpression);
+            report.add_primary(location, file_id, "here".to_string());
+            report
+            }
+            MultiplePragma => {
+                let mut report =
+                Report::error("Multiple pragma directives".to_string(), ReportCode::MultiplePragma);
+            report.add_primary(location, file_id, "here".to_string());
+            report
+            },
+            _ => unreachable!(),    
+    };
+    report
+}
+
